+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - bug fixes
   - remote-root option not applied on resources loaded by the client app (css images, etc)
+  - input: display issue when resizing the window
 
 - editor
   - context menu: add `export` action to export a widget as a fragment file